requirements.txt
# Biomedical Data Platform — Dependencies
# Core
apache-airflow>=2.7.0
pyarrow>=14.0.0
delta-spark>=3.0.0
pyspark>=3.5.0

# FHIR / Healthcare
fhir.resources>=7.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data quality & transformation
great-expectations>=0.18.0
dbt-core>=1.7.0

# Feature store
feast>=0.35.0

# Vector / RAG
sentence-transformers>=2.2.0
qdrant-client>=1.7.0

# Messaging (optional: use kafka-python or confluent-kafka)
confluent-kafka>=2.3.0

# Apps
streamlit>=1.28.0
plotly>=5.0.0
pandas>=2.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0

# Optional: LLM-based FHIR mapping (fhir_omop_agent)
# ollama>=0.1.0  # Uncomment for Ollama/Llama

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
